resave to make minor irrelevant new error message in Max 8.5 go away, reported to Cycling:
Opening the patch on Max 8.5 produces an error msg:

  outlet: patchcord source not found: deleting patchcord

But this doesn’t happen in 8.3 and there’s nothing wrong in the patch.
Resaving it makes the msg go away. The only difference in the file is then this:

 ,       {
           "watchpoint_flags" : 2
-          "watchpoint_id" : 2
-
-,       {
+          "watchpoint_id" : 1

 ,       {

And indeed, there was a watchpoint (very far from the outlet linked to the msg) that changed its number from 2 to 1.